Refactor api-lib.php into smaller, focused libraries in lib/ directory
- Split api-lib.php into lib/env.php, lib/storage.php, lib/business.php, lib/validation.php, lib/audit.php, lib/http.php, lib/email.php
- Preserved api-lib.php as an entry point requiring the new libraries for backward compatibility
- Verified functionality with verify-api-lib.php and curl tests
- Improved code organization and maintainability adhering to SRP
