+++ lib/http.php
@@ -1,0 +1,213 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * HTTP and Session Helpers
+ */
+
+function json_response(array $payload, int $status = 200): void
+{
+    http_response_code($status);
+    header('Content-Type: application/json; charset=utf-8');
+    echo json_encode($payload, JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
+    exit();
+}
+
+function is_https_request(): bool
+{
+    if (isset($_SERVER['HTTPS'])) {
+        $https = strtolower((string) $_SERVER['HTTPS']);
+        if ($https === 'on' || $https === '1') {
+            return true;
+        }
+    }
+
+    if (isset($_SERVER['SERVER_PORT']) && (string) $_SERVER['SERVER_PORT'] === '443') {
+        return true;
+    }
+
+    if (isset($_SERVER['HTTP_X_FORWARDED_PROTO'])) {
+        return strtolower((string) $_SERVER['HTTP_X_FORWARDED_PROTO']) === 'https';
+    }
+
+    return false;
+}
+
+function start_secure_session(): void
+{
+    if (session_status() === PHP_SESSION_ACTIVE) {
+        return;
+    }
+
+    $secure = is_https_request();
+
+    ini_set('session.use_only_cookies', '1');
+    ini_set('session.use_strict_mode', '1');
+    ini_set('session.cookie_httponly', '1');
+    ini_set('session.cookie_secure', $secure ? '1' : '0');
+
+    if (PHP_VERSION_ID >= 70300) {
+        session_set_cookie_params([
+            'lifetime' => 0,
+            'path' => '/',
+            'domain' => '',
+            'secure' => $secure,
+            'httponly' => true,
+            'samesite' => 'Strict'
+        ]);
+    } else {
+        ini_set('session.cookie_samesite', 'Strict');
+        session_set_cookie_params(0, '/; samesite=Strict', '', $secure, true);
+    }
+
+    session_start();
+
+    // Expirar sesion por inactividad
+    if (isset($_SESSION['last_activity']) && (time() - $_SESSION['last_activity']) > SESSION_TIMEOUT) {
+        destroy_secure_session();
+        start_secure_session();
+    }
+    $_SESSION['last_activity'] = time();
+}
+
+function destroy_secure_session(): void
+{
+    if (session_status() !== PHP_SESSION_ACTIVE) {
+        return;
+    }
+
+    $_SESSION = [];
+
+    if (ini_get('session.use_cookies')) {
+        $params = session_get_cookie_params();
+        if (PHP_VERSION_ID >= 70300) {
+            setcookie(session_name(), '', [
+                'expires' => time() - 42000,
+                'path' => $params['path'] ?? '/',
+                'domain' => $params['domain'] ?? '',
+                'secure' => (bool) ($params['secure'] ?? false),
+                'httponly' => (bool) ($params['httponly'] ?? true),
+                'samesite' => 'Strict'
+            ]);
+        } else {
+            setcookie(session_name(), '', time() - 42000, ($params['path'] ?? '/') . '; samesite=Strict', $params['domain'] ?? '', (bool) ($params['secure'] ?? false), (bool) ($params['httponly'] ?? true));
+        }
+    }
+
+    session_destroy();
+}
+
+function require_json_body(): array
+{
+    $raw = file_get_contents('php://input');
+    $data = json_decode($raw ?? '', true);
+    if (!is_array($data)) {
+        json_response([
+            'ok' => false,
+            'error' => 'El JSON enviado no es válido'
+        ], 400);
+    }
+    return $data;
+}
+
+function generate_csrf_token(): string
+{
+    if (!isset($_SESSION['csrf_token']) || $_SESSION['csrf_token'] === '') {
+        try {
+            $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
+        } catch (Throwable $e) {
+            $_SESSION['csrf_token'] = bin2hex((string) microtime(true) . (string) mt_rand());
+        }
+    }
+    return $_SESSION['csrf_token'];
+}
+
+function verify_csrf_token(): bool
+{
+    $token = $_SERVER['HTTP_X_CSRF_TOKEN'] ?? '';
+    if ($token === '' || !isset($_SESSION['csrf_token']) || $_SESSION['csrf_token'] === '') {
+        return false;
+    }
+    return hash_equals($_SESSION['csrf_token'], (string) $token);
+}
+
+function require_csrf(): void
+{
+    if (!verify_csrf_token()) {
+        json_response([
+            'ok' => false,
+            'error' => 'Token CSRF inválido o ausente'
+        ], 403);
+    }
+}
+
+function check_rate_limit(string $action, int $maxRequests = 10, int $windowSeconds = 60): bool
+{
+    $ip = $_SERVER['REMOTE_ADDR'] ?? 'unknown';
+    $key = md5($ip . ':' . $action);
+    $rateDir = data_dir_path() . DIRECTORY_SEPARATOR . 'ratelimit';
+
+    if (!@is_dir($rateDir)) {
+        @mkdir($rateDir, 0775, true);
+    }
+
+    $file = $rateDir . DIRECTORY_SEPARATOR . $key . '.json';
+    $now = time();
+    $entries = [];
+
+    if (file_exists($file)) {
+        $raw = @file_get_contents($file);
+        $entries = is_string($raw) ? (json_decode($raw, true) ?? []) : [];
+    }
+
+    // Filtrar entradas dentro de la ventana de tiempo
+    $entries = array_values(array_filter($entries, static function (int $ts) use ($now, $windowSeconds): bool {
+        return ($now - $ts) < $windowSeconds;
+    }));
+
+    if (count($entries) >= $maxRequests) {
+        return false;
+    }
+
+    $entries[] = $now;
+    @file_put_contents($file, json_encode($entries), LOCK_EX);
+
+    // Limpieza periódica: eliminar archivos de rate limit con más de 1 hora sin modificación
+    if (mt_rand(1, 50) === 1) {
+        $allFiles = @glob($rateDir . DIRECTORY_SEPARATOR . '*.json');
+        if (is_array($allFiles)) {
+            foreach ($allFiles as $f) {
+                if (($now - (int) @filemtime($f)) > 3600) {
+                    @unlink($f);
+                }
+            }
+        }
+    }
+
+    return true;
+}
+
+function require_rate_limit(string $action, int $maxRequests = 10, int $windowSeconds = 60): void
+{
+    if (!check_rate_limit($action, $maxRequests, $windowSeconds)) {
+        json_response([
+            'ok' => false,
+            'error' => 'Demasiadas solicitudes. Intenta de nuevo en unos minutos.'
+        ], 429);
+    }
+}
+
+function verify_admin_password(string $password): bool
+{
+    $hash = getenv(ADMIN_PASSWORD_HASH_ENV);
+    if (is_string($hash) && $hash !== '') {
+        return password_verify($password, $hash);
+    }
+
+    $plain = getenv(ADMIN_PASSWORD_ENV);
+    if (is_string($plain) && $plain !== '') {
+        return hash_equals($plain, $password);
+    }
+
+    return false;
+}
